requirements.txt
fastapi
uvicorn
supabase
pydantic
python-dotenv
qdrant-client
openai
nltk